Extract an `unzipTo` method
